updated model class name and dependencies in controller class
